refactor(ui): consolidate responsive header and search filter components
